[DF] Split test_snapshot into multiple tests to avoid timeout.
